refator: move utility area into utility window directory
