+++ app/chat/components/utilityWindow/utilityArea.tsx
