+++ app/chat/page.tsx
@@ -1,5 +1,5 @@
 import { FriendSideBar } from "@/app/chat/components/friendSideBar";
-import { UtilityArea } from "@/app/chat/components/utilityArea";
+import { UtilityArea } from "@/app/chat/components/utilityWindow/utilityArea";
 import { SidebarProvider } from "@/components/ui/sidebar";
 
 export default async function Page({
